mvp/src/requirements.txt
torch
torchvision
transformers
gradio
anthropic
Pillow
numpy
matplotlib
accelerate